% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target.R
\name{draw_data}
\alias{draw_data}
\title{Freehand drawing}
\usage{
draw_data(data = NULL)
}
\arguments{
\item{data}{Optional \code{data.frame} with x and y values that can used as
background to guide your drawing.}
}
\value{
A \code{data.frame} with the x and y values of your data.
}
\description{
Opens up a dialogue that lets you draw your data.
}
