% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimm.R
\name{keep}
\alias{keep}
\title{Trimm a \code{metamer_list}}
\usage{
keep(object, n = length(object))
}
\arguments{
\item{object}{A \code{metamer_list} object returned by \code{metamerize()}}

\item{n}{The number of metamers to keep.}
}
\value{
A \code{metamer_list} object with \code{n} metamers.
}
\description{
When creating metamers, \code{metamerize()} can produce thousands of very similar
metamers. This function is intended to keep only a subset of them for easier
and faster handling and plotting.
}
