% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mean_dist_to_sf}
\alias{mean_dist_to_sf}
\title{Mean distance to an sf object}
\usage{
mean_dist_to_sf(target, coords = c("x", "y"), buffer = 0, squared = TRUE)
}
\arguments{
\item{target}{An sf object.}

\item{coords}{Character vector with the columns of the data object
that define de coordinates.}

\item{buffer}{Buffer around the sf object. Distances smaller
than \code{buffer} are replaced with 0.}

\item{squared}{Logical indicating whether to compute the mean squared
distance (if \code{TRUE}) or the mean distance.}
}
\description{
Mean distance to an sf object
}
\seealso{
Other helper functions: 
\code{\link{delayed_with}()},
\code{\link{densify}()},
\code{\link{draw_data}()},
\code{\link{mean_dist_to}()},
\code{\link{mean_self_proximity}()},
\code{\link{moments_n}()}
}
\concept{helper functions}
