% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mean_self_proximity}
\alias{mean_self_proximity}
\title{Inverse of the mean self distance}
\usage{
mean_self_proximity(data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
Returns the inverse of the mean minimum distance between differnet pairs of points.
It's intended to be used as a minimizing function to, then, maximize the distance bwetween
points.
}
\seealso{
Other helper functions: \code{\link{delayed_with}},
  \code{\link{densify}}, \code{\link{draw_data}},
  \code{\link{mean_dist_to}}, \code{\link{moments_n}}
}
\concept{helper functions}
