% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamer-class.R
\name{metamer_list}
\alias{metamer_list}
\title{Metamer list object}
\description{
Metamer list object

Metamer list object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{metamer_list$new()}}
\item \href{#method-print}{\code{metamer_list$print()}}
\item \href{#method-minimize}{\code{metamer_list$minimize()}}
\item \href{#method-clear_minimize}{\code{metamer_list$clear_minimize()}}
\item \href{#method-set_minimize}{\code{metamer_list$set_minimize()}}
\item \href{#method-last_metamer}{\code{metamer_list$last_metamer()}}
\item \href{#method-set_annealing}{\code{metamer_list$set_annealing()}}
\item \href{#method-set_perturbation}{\code{metamer_list$set_perturbation()}}
\item \href{#method-set_start_probability}{\code{metamer_list$set_start_probability()}}
\item \href{#method-set_K}{\code{metamer_list$set_K()}}
\item \href{#method-set_change}{\code{metamer_list$set_change()}}
\item \href{#method-as.data.table}{\code{metamer_list$as.data.table()}}
\item \href{#method-as.data.frame}{\code{metamer_list$as.data.frame()}}
\item \href{#method-metamerise}{\code{metamer_list$metamerise()}}
\item \href{#method-metamerize}{\code{metamer_list$metamerize()}}
\item \href{#method-clone}{\code{metamer_list$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new metamer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$new(data, preserve = NULL, signif = 2)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{metamer_list} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-minimize"></a>}}
\if{latex}{\out{\hypertarget{method-minimize}{}}}
\subsection{Method \code{minimize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$minimize(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_minimize"></a>}}
\if{latex}{\out{\hypertarget{method-clear_minimize}{}}}
\subsection{Method \code{clear_minimize()}}{
Removes minimize function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$clear_minimize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_minimize"></a>}}
\if{latex}{\out{\hypertarget{method-set_minimize}{}}}
\subsection{Method \code{set_minimize()}}{
Sets the minimize function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$set_minimize(minimize)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-last_metamer"></a>}}
\if{latex}{\out{\hypertarget{method-last_metamer}{}}}
\subsection{Method \code{last_metamer()}}{
Gets the last metamer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$last_metamer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_annealing"></a>}}
\if{latex}{\out{\hypertarget{method-set_annealing}{}}}
\subsection{Method \code{set_annealing()}}{
Sets \code{annealing}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$set_annealing(annealing)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_perturbation"></a>}}
\if{latex}{\out{\hypertarget{method-set_perturbation}{}}}
\subsection{Method \code{set_perturbation()}}{
Sets \code{perturbation}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$set_perturbation(perturbation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_start_probability"></a>}}
\if{latex}{\out{\hypertarget{method-set_start_probability}{}}}
\subsection{Method \code{set_start_probability()}}{
Sets \code{start_probability}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$set_start_probability(start_probability)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_K"></a>}}
\if{latex}{\out{\hypertarget{method-set_K}{}}}
\subsection{Method \code{set_K()}}{
Sets \code{K}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$set_K(K)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_change"></a>}}
\if{latex}{\out{\hypertarget{method-set_change}{}}}
\subsection{Method \code{set_change()}}{
Sets \code{change}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$set_change(change)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.data.table"></a>}}
\if{latex}{\out{\hypertarget{method-as.data.table}{}}}
\subsection{Method \code{as.data.table()}}{
Gets all metamers as a tidy data.table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$as.data.table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.table}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.data.frame"></a>}}
\if{latex}{\out{\hypertarget{method-as.data.frame}{}}}
\subsection{Method \code{as.data.frame()}}{
Gets all metamers as a tidy data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$as.data.frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.table}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-metamerise"></a>}}
\if{latex}{\out{\hypertarget{method-metamerise}{}}}
\subsection{Method \code{metamerise()}}{
Creates new metamers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$metamerise(
  stop_if = n_tries(10),
  name = "",
  keep = NULL,
  verbose = interactive()
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{metamer_list} object with more metamers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-metamerize"></a>}}
\if{latex}{\out{\hypertarget{method-metamerize}{}}}
\subsection{Method \code{metamerize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$metamerize(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metamer_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
