% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reverse_with}
\alias{reverse_with}
\title{Apply expressions to data.frames}
\usage{
reverse_with(...)
}
\arguments{
\item{...}{Expressions that will be evaluated.}
}
\value{
A function that takes a \code{data.frame} and returns the expressions in \code{...}
evaluated in an environment constructed from it.
}
\description{
Creates a function that evaluates expressions in a future data.frame. Is like
\code{with()}, but you pass the data after.
}
\details{
Each expression in \code{...} must return numeric values. They can be named or
return named vectors.
}
\examples{
some_stats <- reverse_with(mean_x = mean(x), mean(y), sd(x), coef(lm(x ~ y)))
data <- data.frame(x = rnorm(20) , y = rnorm(20))
some_stats(data)

}
