% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{truncate_to}
\alias{truncate_to}
\alias{round_to}
\title{Rounding functions}
\usage{
truncate_to(digits)

round_to(digits)
}
\arguments{
\item{digits}{Number of significant digits.}
}
\description{
Rounding functions
}
\seealso{
Other helper functions: 
\code{\link{delayed_with}()},
\code{\link{densify}()},
\code{\link{draw_data}()},
\code{\link{mean_dist_to_sf}()},
\code{\link{mean_dist_to}()},
\code{\link{mean_self_proximity}()},
\code{\link{moments_n}()}

Other helper functions: 
\code{\link{delayed_with}()},
\code{\link{densify}()},
\code{\link{draw_data}()},
\code{\link{mean_dist_to_sf}()},
\code{\link{mean_dist_to}()},
\code{\link{mean_self_proximity}()},
\code{\link{moments_n}()}
}
\concept{helper functions}
