% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-functions.R
\name{set_minimize}
\alias{set_minimize}
\alias{get_minimize}
\alias{set_preserve}
\alias{get_preserve}
\title{Set \code{metamer_list} attributes}
\usage{
set_minimize(object, minimize)

get_minimize(object)

set_preserve(object, preserve)

get_preserve(object)
}
\arguments{
\item{object}{A \code{metamer_list} object.}

\item{minimize, preserve}{Minimize and preserve functions as defined in
\code{\link[=metamerize]{metamerize()}}.}
}
\description{
Set attributes of \code{metamer_list}s that will be used as function arguments in
\code{\link[=metamerize]{metamerize()}}.
}
