% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-conditions.R
\name{n_tries}
\alias{n_tries}
\alias{n_metamers}
\alias{minimize_ratio}
\title{Stop conditions}
\usage{
n_tries(n)

n_metamers(n)

minimize_ratio(r)
}
\arguments{
\item{n}{integer number of tries or metamers.}

\item{r}{Ratio of minimize value to shoot for. If \code{0.5},
the stop condition is that the iteration will stop if the value
to minimize gets to one-half of the starting value.}
}
\description{
Stop conditions
}
