% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{with_data}
\alias{with_data}
\title{Apply expressions to data.frames}
\usage{
with_data(...)
}
\arguments{
\item{...}{Expressions that will be evaluated.}
}
\value{
A function that takes a \code{data.frame} and returns the expressions in \code{...}
evaluated in an enviroment constructed from it.
}
\description{
Creates a function that evaluates expressions in a future data.frame.
}
\details{
Each expression in \code{...} must return numeric values. They can be named or
return named vectors.
}
\examples{
some_stats <- with_data(mean_x = mean(x), mean(y), sd(x), coef(lm(x ~ y)))
data <- data.frame(x = rnorm(20) , y = rnorm(20))
some_stats(data)

}
