% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-data.R
\name{draw_data}
\alias{draw_data}
\title{Freehand drawing}
\usage{
draw_data(data = NULL)
}
\arguments{
\item{data}{Optional \code{data.frame} with x and y values that can used as
background to guide your drawing.}
}
\value{
A \code{data.frame} with the \code{x} and \code{y} values of your data and a \code{.group} column
that identifies each stroke.
}
\description{
Opens up a dialogue that lets you draw your data.
}
\seealso{
Other helper functions: \code{\link{delayed_with}},
  \code{\link{densify}}, \code{\link{mean_dist_to}},
  \code{\link{moments_n}}
}
\concept{helper functions}
