% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamer-package.R
\docType{package}
\name{metamer-package}
\alias{metamer}
\alias{metamer-package}
\title{metamer: Creates data with identical statistics}
\description{
Creates data with identical statistics (metamers) using an iterative 
   altorithm proposed by Matejka & Fitzmaurice (2017).
}
\section{Overview}{

Create metamers with the \code{\link[=metamerize]{metamerize()}} function. Use the \code{\link[=delayed_with]{delayed_with()}}
to define the statistical transformation you want to preserve.

Inspired by Matejka & Fitzmaurice (2017) awesome paper.
}

\references{
Matejka, J., & Fitzmaurice, G. (2017). Same Stats, Different Graphs. Proceedings of the 2017 CHI Conference on Human Factors in Computing Systems  - CHI ’17, 1290–1294. https://doi.org/10.1145/3025453.3025912
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar}

}
\keyword{internal}
